requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.2
defusedxml==0.7.1
Django==4.1.13
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
idna==3.10
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
oauthlib==3.2.2
psycopg2-binary==2.9.3
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==78.1.0
six==1.17.0
social-auth-app-django==4.0.0
social-auth-core==4.5.6
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.3.0
